+++ problems.php
@@ -163,7 +163,7 @@ class ProblemsPlugin extends Plugin
             $backup_folder = ROOT_DIR . 'backup';
             // try to create backup folder if missing
             if (!file_exists($backup_folder)) {
-                mkdir($backup_folder, 0770);
+                @mkdir($backup_folder, 0770);
             }
         }
 
@@ -172,7 +172,7 @@ class ProblemsPlugin extends Plugin
             $tmp_folder = ROOT_DIR . 'tmp';
             // try to create tmp folder if missing
             if (!file_exists($tmp_folder)) {
-                mkdir($tmp_folder, 0770);
+                @mkdir($tmp_folder, 0770);
             }
         }